pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9"
numpy = "^1.22.2"
bs4 = "^0.0.1"
Flask = "^2.1.2"
pandas = "^1.4.3"
pytz = "^2022.1"
requests = "^2.28.1"
pymongo = "^4.3.3"
discord-py = "^2.1.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"